Fixed box population issues. Began working on basic functionality
Before I was just setting database values to the UI after Query. I fixed it to actually create objects and .Add(object) to the combo box instead.

I set up some conditional selection logic such as in the Create New Database window, when you change the item combo box, it now selects and displays the corresponding price in the window.