Items Commit
Items window is now populated with the items in a gridbox. Users can create items, edit items, and delete items. The gridbox updates when an item is added, edited, or deleted. Users aren't able to change the Item Code of existing items. Attempts to delete items that are in invoices are stopped and the user is warned and informed of which invoices the item is present in.
TODO: Exception handler currently handles cases where the user attempts to add an item with the same code as an existing item (might be fine)
TODO: Closing the items window does not yet update the dropboxes or selected invoice to reflect changes